requirements.txt
twilio
python-dotenv
openai
requests
flask 